Fix transaction feature for all environments (#6)
We initially thought the problem was that the data across each environment was so different that certain customers only existed in certain places.

Actually, the problem was that in some environments the `system` user is set up differently. This means the regimes it has access to is limited, or it was used previously to access a regime other than `CFD` and this has defaulted when the tests run. So, we've made 2 key changes

- switched to a user we know is set up the same across all 3 environments
- updated the test to ensure the right regime is selected before searching

Knowing this we were able to identify a customer that exists across all 3 environments. With these changes plus some going away and creating all the test users across all non-prod environments, all features are now passing everywhere!