Update README intro now all environments covered
With changes we have made to all the non-production environments plus to the features tests themselves, the tests can now be run against any non-production environment and all pass.

There are still only a handful so we need to keep adding to them. But we now have a working base!