Automate test file upload to S3 (#20)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-144

In order to fully test the TCM, we need to be able to upload files to S3 from within our Cypress tests. We implement this by creating a task which reads the bucket and remote path from environment variables CYPRESS_S3_BUCKET and CYPRESS_S3_PATH, and AWS credentials from AWS_REGION, AWS_ACCESS_KEY_ID and AWS_SECRET_ACCESS_KEY. It can then be called from within a test like so:

Given I import the file 'cfdti.dat.csv'
cfdti.dat.csv is an example file added to the Cypress fixtures folder and is used by the import.feature sample test.

Since the task returns a promise, test execution will pause until it either resolves or rejects. If it resolves then the test will continue; if it rejects then the test will fail at that point.

* First pass at upload function

We initially tried putting our upload code in `common/import.js` using AWS SDK v3. We ran into problems importing v3 so we reverted to v2. This at least imported successfully but we were unable to upload files -- it would throw a network error when we tried it. We thought this may be due to `import.js` being run in the Cypress browser so we moved the code into a task `s3Upload` as this would be run by the underlying Node server. And this worked! The next stage will be to see if v3 works in a task, and whether there's a better way of structuring our task (eg. moving it to its own file). But at least we now have a working solution as a base.

* Change credential definition

Since we are now running the AWS SDK in Node, not Cypress, the AWS object can directly access the environment variables. This means we can follow the usual convention of defining `AWS_REGION`, `AWS_ACCESS_KEY_ID` and `AWS_SECRET_ACCESS_KEY` in our environment variables and AWS will pick these up automatically, so we no longer need to use `Cypress.env()` to retrieve them then pass them to the task.

* Param change and small refactor

We change the `Key` param (which is not really the key, as that also includes the filename, but the remote path) to be `fullPath`. We also do a small refactor of how w chain each `.then()`

* Remove commented-out `import` from `import.js`

* Update `.example.env`

* Update aws-sdk to v3

When we were trying to perform the upload in `import.js` we initially had problems with aws-sdk v3 so we reverted to using v2. Once we realised that we had to do the upload in a task, we continued using v2 and developed a working solution. We now replace this with v3 of the sdk.

* Use promise in task

By wrapping our upload command in a promise and returning that promise, Cypress will wait for it to complete before continuing, and if it fails then Cypress will gracefully handle the error.

* Small comment tweak

* Make test pass

We initially added a dummy assertion to our `import.feature` test. To ensure our test successfully passes (even though we have nothing to test!) we simply remove the assertion.