Add @wip tag support (#45)
> This replicates adding @wip support to [sroc-charging-module-api-tests](https://github.com/DEFRA/sroc-charging-module-api-tests/pull/22)

Something we've been able to use in other test projects is the ability to tag 'work in progress' (WIP) scenarios and have the runner only run those. This is immensely helpful when building new tests or trying to figure out why one is failing.

The cypress-cucumber-preprocessor suggests @focus will do the trick but it doesn't seem to work for us. So, by adding additional run scripts to our `package.json` instead, we can mimic the effect.