Add legacy CFD feature from Java project (#41)
https://eaflood.atlassian.net/browse/CMEA-264

The [SROC TCM Acceptance tests](https://github.com/DEFRA/sroc-tcm-acceptance-tests) is actually a single Selenium-based test that goes through the TCM checking it for each of the 3 regimes. It was what the current team were handed and is currently the most extensive test of the TCM service.

However, it is impossible to maintain and hard to follow what the test is actually doing. What we do know of it is that there are no assertions. The test assumes if it can click through with no errors the TCM is functioning correctly. But we have had examples where this is not the case.

This change migrates what the legacy test is doing for the Water Quality (CFD) regime into our Cypress framework. Re-writing the test in BDD format should make it easier to understand what actions are being taken. From there we can review the test, breaking it down into valuable features we can maintain going forward.

** Note - There are some changes from the legacy tests. We have reordered the steps as there was some unnecessary back and forth. Plus we have tried to add more assertions. So, _most_ steps should now have some form of confirmation they have 'worked'.

However, we pretty quickly discarded page objects. This was mainly because we realised what we have needs a good cleanup, including the tests. And we want to bring in the other regimes and then review how best to refactor things and clean up the features.