Automate reading export file (#53)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-282

In `cypress/integration/common/read_export.js` we create an example `When` step which can read a given export file from an S3 bucket, and example `Then` steps which check the value of the item at a given row/column, or check that a value exists in a specified column.

In `cypress/integration/legacy/cfd/cfd_steps.js` we update the `I see confirmation the transaction file is queued for export` step to capture the name of the queued file, and create `I log the transaction filename to prove it can be used in another step` as an example of how we access it from another step.

⚠️ Note that the environment variable `CYPRESS_S3_PATH` (the path we upload import files to) is now changed to `CYPRESS_S3_UPLOAD_PATH` so will need to be updated in existing environment files. We have created an additional `CYPRESS_S3_DOWNLOAD_PATH` which is the path we will download export files from.


* Rename existing S3 env var

We currently have an environment variable `CYPRESS_S3_PATH`, which we use to specify the folder files should be uploaded to. Since we want to be able to specify separate paths for uploading and downloading, we therefore rename this to `CYPRESS_S3_UPLOAD_PATH`

* Create initial `s3Download` task

We start by creatin a basic `s3Download` task which simply reads the specified file and logs its contents to the Cypress console.

* Complete feature

We develop the read export feature to take the file data returned by `s3Download` and split it into a 2-dimensional array. We can then retrieve the item at a given row and column.

* Edit comment

* Write `column contains` step

We write a step which allows us to check that the specified column has a particular value somewhere in it

* Develop legacy cfd steps

In the legacy cfd tests, we update the export file confirmation step to retrieve the filename and store it, and create an additional step which demonstrates how to retrieve the stored filename

* Update `read_export` to use `cy.wrap`

Having realised that `cy.wrap` is preferable to defining variables, we upddate the steps in `read_export`

* Add `CYPRESS_S3_DOWNLOAD_PATH` to `.example.env`