Fix apostrophes for windows (#65)
One of the team who uses Windows ran into a problem when using the **wip** functionality of the project. Even though they were using the tag the relevant scenario was not being run.

After some investigation, we tracked it down to how Windows was handling the call we have in `package.json`. Using double quotes rather than single ones seemed to be a requirement to get it to work on the Windows machine.

So, this change switches the single quotes for double quotes in the package.json. We have tested on our non-windows machines and the calls still work there.