Housekeeping move features to folders (#67)
When we started with Cypress we were still unfamiliar with how we could structure things and therefore put all our feature files at the root of `cypress/integration/`. Since then we've learnt how we can move things into folders and then have multiple feature files grouped around a single domain or feature.

So, we started to use this technique for our later additions but we had not gone back and moved the original tests to this new convention.

This change sorts the issue out.