Add clean DB command (#69)
https://eaflood.atlassian.net/browse/CMEA-281

We recently updated the TCM and [Added a test-only reset DB endpoint](https://github.com/DEFRA/sroc-tcm-admin/pull/544). The purpose was so we could set the database to a known state prior to running certain features in this project.

The legacy features are the key ones; they depend on the database having known records in known states prior to running. This is something we currently do manually but adding a DB clean command is the first step to then automating the legacy tests fully.