Move legacy sort checks to their own feature (#76)
We had started on [Automate db clean and import of regime test files](https://github.com/DEFRA/sroc-acceptance-tests/pull/74) when we started getting some issues with the existing legacy tests. They have always been 'flaky' plus they try and cover so much it can make debugging slow and complicated.

Smack bang in the centre of all 3 is a bunch of steps that check the column sorts in the **Transactions to be billed** screen. In the original legacy tests, we just clicked the column headers. In our legacy tests, we at least attempted to confirm that the sort order was as expected.

But we're still not actually checking that the data is sorted. Plus they make the existing legacy tests more complicated and take longer to run than they need to.

This change moves the sort checking to its own feature with scenarios for each regime. Plus now we can control the data in the system under test we attempt to confirm the data is sorted as expected.