Move legacy transaction detail steps to new scenario (#83)
This is following the same pattern as

- [Move legacy transaction history check to feature](https://github.com/DEFRA/sroc-acceptance-tests/pull/80)
- [Move legacy sort checks to their own feature](https://github.com/DEFRA/sroc-acceptance-tests/pull/76)
- [Move exclude transaction to own feature](https://github.com/DEFRA/sroc-acceptance-tests/pull/78)
- [Move legacy page export steps to new feature](https://github.com/DEFRA/sroc-acceptance-tests/pull/82)

The legacy tests don't do anything more than check that a) you can see the page and b) see the different sections. But it does add noise to the key thing the legacy tests are checking; updating a transaction to the point it can be exported for invoicing.

So, we move them to their own scenario.