Fix step checking for not generated message
The selector for the alert is not the same as the one used for the alert that appears at the top of the page after certain actions.

So, we add a new control to the `ExportDataPage` and use that instead.

Note - We also reduce the amount of text we're checking against. The text often has newlines embedded that are not immediately visible from the UI which can catch you unawares when testing.

As we just want to check the state of the page i.e. that it's reflecting no file has been generated we can simplify the matching by checking for something that still confirms this is the case.