Revert step changes
Completely forgot we have an additional argument we can use to tell the job command not to authenticate when it runs. We added this specifically for this situation; where we are already logged in and need to run the job.

However, this does leave us on the job results page which is why we need to return to the home page before we can then select the regime from the menu.