Downloading a file but is wrong
Have got it to the point that we download the file and save it locally. But it doesn't work. The file we are downloading is a gzip file and what the S3download command is doing is streaming the data as a string.

If we save that to a file any attempts to open it raises a 'unsupported format' error.

Because the cypress tests are run in the browser it has limited access to the file system. You must use it's `writeFile()` method which expects a string.

So, I'm starting to lean towards using the https://www.npmjs.com/package/cypress-downloadfile to get hold of it instead.