Add cypress-downloadfile
This adds support for a `cy.downloadFile()` to the project. It also means we can download the data export in the same way as if we had clicked the download button.

This means we mimic better what a user would do and we avoid the whole issue of managing the data stream from AWS S3 into a format that corecctly copies the file in gzip format.