Check contents of download file
We now have a step that confirms the transaction data file contains the data we expect. It's a cursory check; the number of lines are as expected and each is for a reference that links to the current regime.

But for the purposes of this feature and to confirm the transaction data file export is working this should suffice.